Moved stagefactory to stage directory in main, added blacklist whitelist and textsizelimit config properties
