+++ lucille-core/src/test/java/com/kmwllc/lucille/stage/StageTest.java
@@ -4,7 +4,6 @@ import com.kmwllc.lucille.core.Document;
 import com.kmwllc.lucille.core.Stage;
 import com.kmwllc.lucille.core.StageException;
 import com.kmwllc.lucille.core.UpdateMode;
-import com.kmwllc.lucille.util.StageFactory;
 import com.typesafe.config.Config;
 import org.junit.Test;
